pyproject.toml
[tool.poetry]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = ["michaelockenden <michael.ockenden@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websocket-client = "^1.7.0"
rel = "^0.4.9.8"
pandas = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
